requirements.txt
playwright
opencv-python
numpy
pyautogui
pywin32; sys_platform == 'win32'
Pillow